# application/services/custom_services/testapp/src/models/entitymodels/test.ts
export interface test 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: String,
   email: String,
   test: String
}